Refactored and new story
